[Refactor] utils쪽 함수들 다시 병합
